backend/requirements.txt
asgiref==3.8.1
Django==5.0.2
djangorestframework==3.14.0
python-decouple==3.8
sqlparse==0.5.3
django-cors-headers==4.3.1
dj-database-url==2.1.0
djoser==2.2.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
psycopg2-binary==2.9.9